fix: 동시성 이슈 처리
- 꾸러미 할당 시의 동시성 이슈를 처리하기 위한 클래스를 추가하였습니다.
